Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
celery==5.3.4
loguru==0.7.2
prometheus-client==0.19.0
slowapi==0.1.9
aiocache==0.12.2
websockets==12.0
python-socketio==5.10.0
email-validator==2.1.0
Pillow==10.1.0
boto3==1.34.0
cryptography>=41.0.0
schedule==1.2.0 